backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
sqlalchemy==2.0.36
pypdf==5.1.0
python-docx==1.1.2
openai==1.57.0
httpx==0.28.1
pydantic-settings==2.6.1
email-validator==2.2.0
